// api-portal/src/types/registry.d.ts
export interface operations {
    "specification/get/specifications": {
        parameters: {
            query?: {
                fields?: ("specId" | "name" | "productName" | "version" | "revision" | "bundleName" | "componentName" | "componentRevision" | "contract" | "body" | "score" | "violations" | "createdAt" | "updatedAt" | "publishedAt")[];
                filters?: {
                    [key: string]: unknown;
                };
                _q?: string;
                sort?: ("specId" | "name" | "productName" | "version" | "revision" | "bundleName" | "componentName" | "componentRevision" | "contract" | "body" | "score" | "violations" | "createdAt" | "updatedAt" | "publishedAt")[];
                populate?: never[];
                status?: "draft" | "published";
                publicationFilter?: "never-published" | "has-published-version" | "modified" | "unmodified" | "never-published-document" | "has-published-version-document" | "published-without-draft" | "published-with-draft";
                hasPublishedVersion?: "true" | "false";
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        data: {
                            documentId: string;
                            id: string;
                            specId: string;
                            name: string;
                            productName: string;
                            version: string;
                            revision: string;
                            bundleName: string;
                            componentName?: string;
                            componentRevision?: string;
                            /** @enum {string} */
                            contract: "OPENAPI_V30" | "OPENAPI_V31" | "OPENAPI_V32" | "ASYNCAPI_V2" | "ASYNCAPI_V3" | "GRAPHQLS" | "WSDL_V11" | "WSDL_V20" | "OVERLAY_V10" | "OVERLAY_V11";
                            body: unknown;
                            score?: unknown;
                            violations?: unknown;
                            createdAt?: string;
                            updatedAt?: string;
                            /** @default 2026-07-18T05:31:36.588Z */
                            publishedAt: string;
                        }[];
                    };
                };
            };
            /** @description Bad request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Unauthorized */
            401: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Forbidden */
            403: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Not found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Internal server error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
// ...
}